[REFACTORING] Split IMAPServerConnectionCheckTest
